Apply improved types & architecture to Cardano
