+++ libs/ledger-live-common/src/families/cardano/bridge/js.ts
@@ -1,10 +1,11 @@
 import type { Transaction } from "../types";
-import { SignerContext, scanAccounts, sync } from "../js-synchronisation";
-import estimateMaxSpendable from "../js-estimateMaxSpendable";
-import { createTransaction, prepareTransaction } from "../js-transaction";
-import getTransactionStatus from "../js-getTransactionStatus";
-import signOperation from "../js-signOperation";
-import broadcast from "../js-broadcast";
+import { SignerContext, scanAccounts, sync } from "../synchronisation";
+import estimateMaxSpendable from "../estimateMaxSpendable";
+import { prepareTransaction } from "../prepareTransaction";
+import { createTransaction } from "../createTransaction";
+import getTransactionStatus from "../getTransactionStatus";
+import signOperation from "../signOperation";
+import broadcast from "../broadcast";
 import { makeAccountBridgeReceive } from "../../../bridge/jsHelpers";
 import { defaultUpdateTransaction } from "@ledgerhq/coin-framework/bridge/jsHelpers";
 import type { AccountBridge, CurrencyBridge } from "@ledgerhq/types-live";
