+++ libs/ledger-live-common/src/families/cardano/broadcast.ts
@@ -1,12 +1,13 @@
-import type { BroadcastFnSignature } from "@ledgerhq/types-live";
+import { AccountBridge } from "@ledgerhq/types-live";
 import { patchOperationWithHash } from "../../operation";
 import { submitTransaction } from "./api/submitTransaction";
+import { Transaction } from "./types";
 
 /**
  * Broadcast the signed transaction
  * @param {signature: string, operation: string} signedOperation
  */
-const broadcast: BroadcastFnSignature = async ({ signedOperation, account }) => {
+const broadcast: AccountBridge<Transaction>["broadcast"] = async ({ signedOperation, account }) => {
   const signedTx = signedOperation.signature;
   const pendingTransaction = await submitTransaction({
     transaction: signedTx,
