+++ libs/ledger-live-common/src/families/cardano/estimateMaxSpendable.ts
@@ -1,30 +1,26 @@
-import { BigNumber } from "bignumber.js";
-import { log } from "@ledgerhq/logs";
-import type { AccountLike, Account } from "@ledgerhq/types-live";
-import { getMainAccount } from "../../account";
-import type { CardanoAccount, Transaction } from "./types";
-import { createTransaction } from "./js-transaction";
 import {
   address as TyphonAddress,
   types as TyphonTypes,
   Transaction as TyphonTransaction,
 } from "@stricahq/typhonjs";
-import { buildTransaction } from "./js-buildTransaction";
+import { BigNumber } from "bignumber.js";
+import { log } from "@ledgerhq/logs";
+import type { AccountBridge } from "@ledgerhq/types-live";
+import type { CardanoAccount, Transaction } from "./types";
+import { createTransaction } from "./createTransaction";
+import { buildTransaction } from "./buildTransaction";
+import { getMainAccount } from "../../account";
 
 /**
  * Returns the maximum possible amount for transaction
  *
  * @param {Object} param - the account, parentAccount and transaction
  */
-const estimateMaxSpendable = async ({
+const estimateMaxSpendable: AccountBridge<Transaction>["estimateMaxSpendable"] = async ({
   account,
   parentAccount,
   transaction,
-}: {
-  account: AccountLike;
-  parentAccount?: Account;
-  transaction?: Transaction;
-}): Promise<BigNumber> => {
+}) => {
   const mainAccount = getMainAccount(account, parentAccount);
 
   if ((mainAccount as CardanoAccount).cardanoResources.utxos.length === 0) {
@@ -38,7 +34,7 @@ const estimateMaxSpendable = async ({
   const dummyRecipient =
     "addr1qyqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqv2t5am";
   const t: Transaction = {
-    ...createTransaction(),
+    ...createTransaction(account),
     ...transaction,
     recipient: dummyRecipient,
     // amount field will not be used to build a transaction when useAllAmount is true
